Getting docker working after grpc added
This required (unfortunately) at least for now to move to an
Ubuntu image instead of alpine. This is becasue the grpcio-tools
and grpcio python libs use cython which in turn has some stronger
dependencies on runtime packages, and we could not yet resolve
all the issues arising on alpine.

Changes:
- committed *.desc and *_pb2.py output files so that protoc tools
  are not needed when building dockerimage
- moved google protoc extension files (http.proto and annotations.proto
  to under python source tree. Otherwise they were not copied into
  the docker image
- created Ubuntu Dockerfile while we retained the old apline file
  so that we can resurrect the alpine version later