Initial take on persistence support via k/v stores,
This is the initial mechanisms for persistence, supported
by a k/v store. In the initial testing the k/v store
is mcoked by a simple (in-memory) dict, but the concept
will carry to external, replicated, distributed k/v
store options, such as cassandra, redis, or even
consul for small-scale deployments.

Change-Id: I83d2bbb327c4516fbc15d1d9979a1e89d3e7a7e7