Define and wire in bulk flow update adapter API
This adds support for defining bulk or incremental
flow update capability on a per device type basis
and implements the callout mechanism.

It also defines 5 new API entries for proxying
messages from a child-device adapter (e.g. ONU adapter)
to the actual device as well as the APIs for registering
and receiving async messages from a device via a
parent device acting as a proxy. The implementation
of this is left for the next PR.

Change-Id: Ic48a458c170083842b6bc674d675b5b60c0827f6