preparing to split chameleon out of voltha
this is a prep step to make chameleon independent
of voltha. Chameleon will move out of voltha but
repo will drop chameleon in the same place as it
currently is. This means that voltha's build process
will not change and the code changes to chameleon
will automatically be applied to the correct repo.

Change-Id: I754d6b5b28ea99333b19140d6c1a94e8198f9d3a