Inject per-ONU metadata field for unicast flows
This is a CLI change to mimic a useful ONOS behavior when generating
logical flows for the PON. Specifically, ONOS injects a metadata
field in each flow rule for unicast downstream traffic, namely into
the first of the two flow rules handling the outer tag. The metadata
value is the vlan id of the inner tag. Without this metadata there
is no easily accessible information as to what inner tag that flow
is meant for.

This metadata value can be considered as a "hint" by the OLT adapters
to tie a downstream flow rule to a specific PON link/channel.

This is not an elegant solution, in that it slightly misuses the
metadata field. The more proper long-term solution would be to either
model the PON channels explicitly as flow ports, or use phys-port/port
pairs (the former representing the PON port itself, and the other
representing the logical channel/link on the PON.

It is recommended to switch to the cleaner solution at a later time.

Change-Id: I2a461014d697d01010101010101052609d742d04