Add consul as a config backend
Two backends available, none and consul. Can be set through the
BACKEND environment variable or the --backend flag.
Currently access to consul blocks the reactor. Requests to consul
are cached. In my conversations with Zsolt, we concluded that
blocking the reactor is fine for now to prevent multiple tasks
from modifying the config at the same time. An easier first step
to making it all twisted is to add a write queue. That comes with its
own challenges though. A longer goal would be making the entire config
access inlineCallbacks, possibly with a locking mechanism to prevent
multiple tasks from modifying the config at once.

Change-Id: Ia3132a455ca03123dbb9b2e56c9f4df1202347cd