VOL-217 Initial implementation of device handler for asfvolt16
This commit is a copy-paste of the maple_olt device handler.
This is not elegant but it is being pushed to the repo to
enable others to start integrating with the gRPC protobuf
backend. Subsequent commits will introduce a base device
handler class which will abstract out the generic functions.

Change-Id: I7cfb25ece6cce0ab399d4b4083593ff9215c5270