VOL-211: Integrate a clustered Consul service with all Voltha components that use it
The Dockerfile for the CLI container runs a setup.sh script at container startup.
This script hard-codes the Consul lookup (-L) and endpoint (-C) arguments passed
to /cli/cli/main.py. In Docker swarm mode, Consul is no longer used for service
lookup; instead, Docker's DNS is used. It is desirable that the lookup argument
be specified in the compose file. For v2 -L is supplied and the Voltha gRPC and
SIM endpoints are specified by Consul service name. For v3 -L is withheld and the
Voltha gRPC and SIM endpoints are specified by Docker service name. Therefore, in
addition to a new v3 compose file for use by Docker swarm, the existing v2 compose
file used by the single-node Voltha implementation must be changed.

Change-Id: Ifef3e06ef9479594bfb11328dd1b1004e4f6c74f