VOL-282: Integrate demo components (e.g. dashd, shovel) with the clustered consul service
This submission migrates the dashd service to Docker swarm mode. The dashd implementation
did not use the '@' character to distinguish between a Consul service and a Docker service.

This submission adds support for both service types and a v3 compose file for a 'stats'
service stack. The shovel service will be added to this stackfile in a future submission.

Change-Id: Iad4695641d489666ed27a78b0a79539357a8e2be