[4647] VOL-282: Integrate demo components (e.g. dashd, shovel) with the clustered consul service
The grafana Dockerfile was modified to facilitate interworking with nginx. The only
modification made beyond the base kamon/grafana_graphite image was to the root_url
property in file /opt/grafana/conf/custom.ini. Reverting to the base image removes
the nginx requirement and allows access to the grafana UI via port 8882.
The v3 compose file for nginx is removed from the repo.

Change-Id: Ic3735487713287af1555f902bffec331950dfc9d