VOL-580: Update envoy to support either consul or etcd in a kubernetes cluster
Envoy is refactored such that it interworks with either consul or etcd. The refactoring
should not have changed any of envoy's behaviour with consul. The update was tested with
only the vcli, envoy, vcore, and KV-store (consul or etcd) containers:
- tested with consul on a kubernetes 3-node cluster
- tested with etcd on a kubernetes 3-node cluster
- tested with consul on a 3-node docker swarm
- container and node failure testing not performed

This update enables some experimentation of envoy in the kubernetes environment.
There is more work to be done.  Future updates could:
- replace the current polling of etcd's assignment key with an asynchronous technique
- make use of data consistency options in etcd get requests
- make greater use of object orientation

Change-Id: Ia22893cdf331967eee4d13e060fd2c5ba7adb321