VOL-569: Create kubernetes deployment configuration for each voltha service
This update adds a kubernetes deployment file for netconf.

I tested the netconf service using the CLI that comes with the netopeer container.
That CLI appears to make connections with remote netconf servers at port 830; that
port number appears to be hard-coded. The port of the netconf server, on the other
hand, defaults to 1830. The changes in file netconf/main.py enable the override of
this default by using environment variable NETCONF_PORT. The kubernetes deployment
file for netconf sets the port to 830.

Change-Id: I19f90d6e14f59c855b84aa76bbc41586f3bace37