[8055] This update adds kubernetes deployment files for the OFagent and fluentd services.
The OFagent calls get_my_containers_name with the assumption that the HOSTNAME
environment variable contains the docker container's ID, which does not appear
to be the case for kuberbetes. In kubernetes a container's HOSTNAME is its pod
name. Feeding a kubernetes pod name into the docker API used by
get_my_containers_name triggers an exception. This is the reason behind the
code changes to OFagent.

The kubernetes config files in this submission still use the old image naming
convention. The names in these and all the other kubernetes deployment files
will be changed to conform to the new convention in a separate update.

Change-Id: I3eecfdcc8bba070ffe3e7372f195cf15f0cdcd56