VOL-572: Integration testing with Kubernetes
This update is an attempt to modify the Voltha integration test suite to support
multiple run-time environments: single-node Kubernetes, single-node Docker swarm,
as well as the current docker-compose environment. To run tests in environments
other than docker-compose, a config file containing test parameters is passed to
the test code via a nose plugin call nose-testconfig. The path to this file is
specified via a "tc-file" argument supplied to the nosetests command.

Thus far, only test_cold_activation_sequence and test_voltha_rest_apis have been
modified. The intent is to update the remaining integration tests as well. The
README.md file has been updated for these 2 tests but this is not necessarily
how the documentation will evolve with this feature.

Change-Id: I6d9b260c34ef069935ae30958f3c3012ffe603b6