VOL-572: Integration testing with Kubernetes
Updated test_dispatcher to run in the single-node Kubernetes environment, as well as
in docker-compose.

Test_dispatcher.py requires the 'scenario' object literal defined in test_voltha_xpon.py,
which it imports from that file. The import operation appears to cause code in test_voltha_xpon.py
to execute, code which requires containers to be already running. This defeats the automation
that was already built into test_dispatcher by forcing the user to manually deploy containers.
This update removes 'scenario' from test_voltha_xpon.py and puts it in a separate file, which
is then imported by each of these tests.

Change-Id: Ia049ae44686358606939daceab6543e9d455c261