VOL-950: Lookup onu device based on PON-intf_id and onu id
The adapter_agent.get_child_device() method is used by OLT adapters
to lookup the ONU device. The lookup allows search on ONU ID
(and/or ONU serial number).
When looking up by ONU ID, this method should also take the
PON interface id into consideration since ONU IDs are not
unique across PONs.

The parent_port_no member of the Device class is used to identify
the PON interface that the ONU is connected to.

Change-Id: I19478e73157c74e095fb84b327b317539fc10276