VOL-1030 Ignore bulk flow update if OLT is operational down.
This seems to fix voltha from 'hanging' when the openolt
driver is killed. Not entirely certain why - but it appears
that, when the gRPC connection is lost, the logical device's
state is made to be down. This causes ONOS to delete flows
trigerring the bulk flow update to the adapter. The adapter
attempts to make gRPC calls to the driver which fail after
connection timeouts. Likely twisted thread is busy with
the flow programming causing voltha to hang.

Change-Id: I9f07479f85ce2bf6cffe40c663581c013c897495