VOL-1180: Enable ponsim to send packet-outs on NNI port.
This required adding the out_port metatdata to the message that
VOLTHA uses to send packets to ponsim, so that VOLTHA can tell
ponsim which port to send the packet out.

Also refactored packet-outs in ponsim so that they don't get run
through the forwarding pipeline, they just get sent out the port.
This is what happens in a normal openflow device.

Change-Id: Ic40ea730f061d82d2ecee33fb1a1f371d0ba73ef