VOL-1289 - Update Resource Manager to support per-interface ranges
Updated Resource Manager and OpenOLT to support constructs for managing
resource pools by 1) each interface, 2) by all interfaces that share
a common technology, and 3) by all interfaces on the board.

The OpenOLT agent now supports returning, via DeviceInfo, a list of
1 or more range specifications. Each range corresponds to a unique
technology and identifies which interfaces on the board use that
technology and the start/end specifications for each resource type,
e.g. resource "pool". Options are provided for each pool to
independently be configured for (1) pool-per-interface, (2) pool-
per-technology, or (3) pool-per-device, a.k.a global.

A separate Resource Manager instance is used for each technology
range specification that manages (1) and (2) options. For (3),
one of the Resource Managers (arbitrarily chosen) is designated
as the "global" Resource Manager that is delegated to from
the subordinate RMs for those pools configured for (3) pool-per-
device.

For all pools, the Tech Profile is first read, if present. Then,
the device's technology range specification is overlayed to form a
subset of overlapping ranges. Then, globally-shared pools overalyed
again to form the overlapping ranges amongst all learned DeviceInfo
pools and TechProfile ranges.

Presently, behavior is undefined if no overlapping range exists
between TP and DevInfo ranges.

Backwards compatibility with OpenOLT agent drivers that only
support the original, single technology interface by synthesizing
a single range encompassing all interfaces with the start/end
specifications. In this case, the EdgeCore/BAL limitation for
resource sharing is set for Alloc ID and GEM Port ID.

Change-Id: I507ac013a114950dc0155b98a38406a42e2fba87