VOL-1311 - OpenOMCI read failure when "get" action on ME table attribute
Temporary work around in brcm_openomci_onu adapter to avoid get of
table attribute in ExtendedVlanTaggingConfigurationData ME that occurs
implicitly after create if the VlanTaggingTable attribute was not
set during creation. This is causing exception on some ONUs during
MIB download.

Workaround is to make sure to set the VlanTaggingTable attribute
during the create operation rather than as a subsequent set operation.

Change-Id: I44a9151bdbdb2febb5b12f97ae7cd395eccdb74e