VOL-1322 - EAPOL messages dropped before reaching OLT
Partially reverted VOL-1311. ExtendedVLANTaggingOperationConfigurationData
cannot have the VLAN Table set during create as it is a R/W attribute. However, it must
be set per VOL-1311 during create to work around a parsing error in a get
of missing W attributes not present in the create. Current approach
is to set the table as a bogus set in the create() and then set again
immediately afterwards in a set().

Change-Id: I251fc177fd1bb49d1bc69cba9c8c7688ad88e674