VOL-1333 - OpenOMCI exception caused by missing set-table to ONU and read-table from DB
Added support to complete the set transaction for an OMCI table attribute. If using
the OmcitableField, the omci will automatically handle converting a set of a single row
and upon completion of the set, cause a table "update" to occur to augment the new
row into the existing table. The supplier must provide an index() and is_delete() method
to support determination if the row set() is updating an existing row, deleting
an existing row, or adding a  new row. Rows are sorted by index() order.

The ramification is also a change in the query() contract for a table attribute. It now
returns a list[] of objects rather than a single scaler object. Presently, this
only affects query of the ExtendedVlanTaggingOperationData table.

Change-Id: I2b24f747beb79013f078bbb8c37006e75fda0712