VOL-1343 Disable TLS in ofagent compose and onos Docker build
This change allows onos development builds to be more easily tested
in a wider range of environments.  The base onosproject/onos build
does not have TLS certs built in.  Now that the voltha dev build
does not it can be swapped in places where onosproject/onos is running
without TLS mismatch configuration problems.

Also update the docker-compose dev setup scripts for ofagent to disable TLS,
again to be consistent with onos and helm deployed versions of ofagent.

Change-Id: I8a3d2c02e7091200bd318edd170c1fb26d34e068