VOL-1438 SEBA-344 Force mib reset on device re-enable
Workaround related to mib sync issues.  Given the omci library
now checks for MDS value on start (or restart) it skips the mib
reset needed to clear out of vlan config data.  Override the startup
method to unset the last mib sync value and forcing a resync and mib
upload

Change-Id: I07e47af56f219fb1f58cd0d310a377c7118c71fa