VOL-1352 - Use serial_number reported by device instead of host:port.
Currently the logical device's serial number is set to the
host:port of the olt. This was done since the device's
serial number was not reported by the openolt agent.
Since the openolt agent now reports the actual device serial
number, this commit makes it a requirement to use the
device's serial number in the onos config.

This change will require onos sadis config to specify
the serial number instead of host:port:

                "entries" : [
                    {
                        "id" : "EC1721000216",
                        "hardwareIdentifier" : "de:ad:be:ef:ba:11",
                        "uplinkPort" : 65536
                    },

This commit requires the corresponsing openolt agent change which
provides the serial number in the device_info:

commit 42bc6ec6af647ebd42e690c0e28e1d5623ab912f
Author: Thiyagarajan Subramani <Thiyagarajan.Subramani@radisys.com>
Date:   Sat Feb 2 03:21:43 2019 -0800

    VOL-1392: OpenOLT driver should send the actual device serial number

    Change-Id: I1c9703568bc85f7e8e3c62313a4a9abaa9d7b1e7

Change-Id: I9a40717baf6ca23d6a1171d4e79f49a0c5175133