Revert assert removed in previous commit
The previous commit removed a call to resource_mgr
that would assert if (uni_id != 0).
This commit adds that assert back (but as a local assert, not a
call to resource_mgr).
At this time multiple unis per onu is not supported.
This assert should be removed once multi unis are supported.

Change-Id: Id5b1f04cc7469a9979fa3fb73ba767c03a3fc5b2